+++ rocketgram/executors/tornado.py
@@ -6,14 +6,14 @@
 import asyncio
 import logging
 import signal
-from typing import TYPE_CHECKING, Union, Dict, List, Set
+from typing import TYPE_CHECKING, Union, Dict, List, Set, Optional
 
 from tornado.httpserver import HTTPServer
 from tornado.httputil import HTTPServerRequest, HTTPHeaders, ResponseStartLine
 
 from .executor import Executor
 from ..api import Request, GetMe, SetWebhook, DeleteWebhook
-from ..api import Update
+from ..api import Update, UpdateType
 from ..errors import RocketgramRequestError
 from ..version import version
 
@@ -32,8 +32,7 @@ HEADERS_ERROR = {"Server": f"Rocketgram/{version()}", "Content-Type": "text/plai
 
 
 class TornadoExecutor(Executor):
-    def __init__(self, base_url: str, base_path: str, *, host: str = 'localhost',
-                 port: int = 8080):
+    def __init__(self, base_url: str, base_path: str, *, host: str = 'localhost', port: int = 8080):
         """
 
         :param base_url:
@@ -72,14 +71,16 @@ class TornadoExecutor(Executor):
     def can_process_webhook_request(self, request: Request) -> bool:
         return len(request.files()) == 0
 
-    async def add_bot(self, bot: 'Bot', *, drop_pending_updates=False, suffix=None, set_webhook=True,
-                      max_connections=None):
+    async def add_bot(self, bot: 'Bot', *, allowed_updates: Optional[List[UpdateType]] = None,
+                      drop_pending_updates: bool = False, suffix: str = None, set_webhook: bool = True,
+                      max_connections: int = None):
         """
 
         :param bot:
+        :param allowed_updates:
+        :param drop_pending_updates:
         :param suffix:
         :param set_webhook:
-        :param drop_pending_updates:
         :param max_connections:
         """
         if bot in self.__bots.values():
@@ -102,8 +103,9 @@ class TornadoExecutor(Executor):
         logger.info('Added bot @%s', bot.name)
 
         if set_webhook or drop_pending_updates:
-            await bot.send(SetWebhook(full_url, drop_pending_updates=drop_pending_updates,
-                                      max_connections=max_connections))
+            swh = SetWebhook(full_url, allowed_updates=allowed_updates, drop_pending_updates=drop_pending_updates,
+                             max_connections=max_connections)
+            await bot.send(swh)
             logger.debug('Webhook setup done for bot @%s', bot.name)
 
         if drop_pending_updates:
@@ -247,9 +249,10 @@ class TornadoExecutor(Executor):
         logger.info("Stopped.")
 
     @classmethod
-    def run(cls, bots: Union['Bot', List['Bot']], base_url: str, base_path: str, *, host='localhost', port=8080,
-            webhook_setup=True, webhook_remove=True, drop_pending_updates=False,
-            signals: tuple = (signal.SIGINT, signal.SIGTERM), shutdown_wait=10):
+    def run(cls, bots: Union['Bot', List['Bot']], base_url: str, base_path: str, *, host: str = 'localhost',
+            port: int = 8080, webhook_setup: bool = True, webhook_remove: bool = True,
+            allowed_updates: Optional[List[UpdateType]] = None, drop_pending_updates: bool = False,
+            signals: tuple = (signal.SIGINT, signal.SIGTERM), shutdown_wait: int = 10):
         """
 
         :param bots:
@@ -259,6 +262,7 @@ class TornadoExecutor(Executor):
         :param port:
         :param webhook_setup:
         :param webhook_remove:
+        :param allowed_updates:
         :param drop_pending_updates:
         :param signals:
         :param shutdown_wait:
@@ -269,7 +273,8 @@ class TornadoExecutor(Executor):
         executor = cls(base_url, base_path, host=host, port=port)
 
         def add(bot: 'Bot'):
-            return executor.add_bot(bot, drop_pending_updates=drop_pending_updates, set_webhook=webhook_setup)
+            return executor.add_bot(bot, allowed_updates=allowed_updates, drop_pending_updates=drop_pending_updates,
+                                    set_webhook=webhook_setup)
 
         def remove(bot: 'Bot'):
             return executor.remove_bot(bot, delete_webhook=webhook_remove)